Once upon a time, Apple was so concerned about accuracy of floating point operations that they required all operations be performed in IEEE extended (80bit). This had the interesting side effect that a program which conformed to Apple's standards could not also conform to ANSI C (which required operations use double temporaries for double values). I find it somewhat ironic, therefore, that gcc on the mac doesn't even promote floats to doubles. FF needs the extra accuracy provided by a double (an extended type doesn't seem to provide additional assistance unfortunately). So on the mac we got extrema that weren't, which meant that remove overlap didn't.

+++ fontforge/anchorsaway.c
@@ -438,10 +438,10 @@ static void AnchorD_DrawPos(AnchorDlg *a) {
     char buffer[20];
     unichar_t ubuf[20];
 
-    sprintf( buffer, "%g", a->apos.x );
+    sprintf( buffer, "%g", (double) a->apos.x );
     uc_strcpy(ubuf,buffer);
     GGadgetSetTitle(GWidgetGetControl(a->gw,CID_X),ubuf);
-    sprintf( buffer, "%g", a->apos.y );
+    sprintf( buffer, "%g", (double) a->apos.y );
     uc_strcpy(ubuf,buffer);
     GGadgetSetTitle(GWidgetGetControl(a->gw,CID_Y),ubuf);
